feat: split up display into separate functions
