+++ README.md
@@ -4,27 +4,38 @@ Legacy support for our utility scripts implementing async, functional, and immut
 
 Previously, this project was in separate repositories for:
 
-- display
-  - Display
+- Display
+  - `coerceRealNumber`
+  - `decimalPlaces`
+  - `formatBytes`
+  - `formatCurrency`
+  - `formatDuration`
+  - `formatNumber`
+  - `humanizeDuration`
+  - `humanizeSessionDuration`
+  - `numberWithCommas`
+  - `prefixWith`
+  - `suffixWith`
+  - `unitConversion`
 - functional-util
 - maybe
-  - Maybe
+  - `Maybe`
 - immutable-optics
-  - traversal
-  - postOrderTraversal
-  - preOrderTraversal
+  - `traversal`
+  - `postOrderTraversal`
+  - `preOrderTraversal`
 - immutable-util
-  - immMemoize
-  - list
-  - walk
-  - patch
-  - remove
-  - startsWith
-  - endsWith
-  - isInstance
-  - ImmutableRecordMemoize
-  - cartesianProduct
-  - transpose
-  - complement
-  - splitIntoChunks
-  - partitionB
+  - `immMemoize`
+  - `list`
+  - `walk`
+  - `patch`
+  - `remove`
+  - `startsWith`
+  - `endsWith`
+  - `isInstance`
+  - `ImmutableRecordMemoize`
+  - `cartesianProduct`
+  - `transpose`
+  - `complement`
+  - `splitIntoChunks`
+  - `partitionB`
