+++ decimalPlaces.js
@@ -1,0 +1,7 @@
+module.exports = function decimalPlaces(digits) {
+    return (numberArg) => {
+        const number = parseFloat(numberArg);
+        const multiplyBy = Math.pow(10, digits);
+        return (Math.round(number * multiplyBy) / multiplyBy).toFixed(digits);
+    };
+}
