+++ index.js
@@ -1,11 +1,9 @@
-const Display = require('./display.js');
 const Maybe = require('./maybe.js');
 const functionalUtil = require('./functional-util.js');
 const immutableOptics = require('./immutable-optics.js');
 const immutableUtil = require('./immutable-util.js');
 
 module.exports = {
-  Display,
   Maybe,
   functionalUtil,
   immutableOptics,
