+++ package.json
@@ -4,12 +4,23 @@
   "description": "Legacy support for our utility scripts implementing async, functional, and immutable JavaScript.",
   "main": "index.js",
   "files": [
-    "display.js",
+    "coerceRealNumber.js",
+    "decimalPlaces.js",
+    "formatBytes.js",
+    "formatCurrency.js",
+    "formatDuration.js",
+    "formatNumber.js",
     "functional-util.js",
+    "humanizeDuration.js",
+    "humanizeSessionDuration.js",
     "immutable-optics.js",
     "immutable-util.js",
     "index.js",
     "maybe.js",
+    "numberWithCommas.js",
+    "prefixWith.js",
+    "suffixWith.js",
+    "unitConversion.js",
     "README.md"
   ],
   "author": "Quantum Metric",
@@ -35,10 +46,10 @@
     "@semantic-release/changelog": "^6.0.1",
     "@semantic-release/git": "^10.0.1",
     "chai": "^5.1.0",
-    "mocha": "^10.1.0",
-    "nyc": "^17.0.0",
-    "semantic-release": "^24.0.0",
-    "standard": "^17.0.0"
+    "mocha": "^10.7.3",
+    "nyc": "^17.1.0",
+    "semantic-release": "^24.1.1",
+    "standard": "^17.1.2"
   },
   "dependencies": {
     "immutable": "^3.8.2",
