+++ prefixWith.js
@@ -1,0 +1,3 @@
+module.exports = function prefixWith(prefix) {
+    return (x) => `${prefix}${x}`;
+}
