+++ suffixWith.js
@@ -1,0 +1,3 @@
+module.exports = function suffixWith(suffix) {
+    return (x) => `${x}${suffix}`;
+}
